Add LTE::Game, replace LTE::Player with the more generic LTE::Character
There were certain aspects of the game-state that didn't really make sense to
mash into the player class, like time of day, current location, global events,
etc... so a game class was created (similar to the earlier LTE::Profile) which
took the place of player in the cases where the game-state was loaded and saved.
LTE::Game#player is of class character, and takes the place of LTE::Player in
all other cases.

LTE::Player was removed because once the save/load functionality was removed, it
became identical to LTE::Character.