Move some logic from bin/lte to resources/lte.glade
When the userdata for a callback is set in the glade file, the ruby gtk builder
will try to call the callback method as an instance method of the userdata
object. For example, the help_about_menuitem's active callback is set to 'show'
with the about_dialog window in the userdata field, so when the builder is
connecting signals it will connect help_about_menuitem -> active to
about_dialog.show. This was used to remove useless one-line callbacks like
the help_about_menuitem's activate callback, which just showed the about_dialog
window.

When there was extra logic in the menuitem's activate callback (such as in the
case of game_save_as_menuitem, where it needs to check if a @game object exists
before it can save it), the code was moved to the show callback for the
respective window (in the example's case, game_save_filechooser_dialog_show).
This makes the code more general, as you don't have to reimplement it everywhere
you want to show the window. It is done automatically when the window is shown.

The game_save_menuitem's activate callback was kept, because it could either
save the game or (if there was no file associated with the @game object) open
the save filechooser. There is no window show callback where the code to check
if the @game_save_file existed could be moved, so it was kept in the
game_save_menuitem_activate callback.